chore: lift authenticated app RootProvider into AppContainer
Move the authenticated app RootProvider from AppMain to AppContainer so AppMain, global modals, calling audio wrappers, and normal authenticated app content share the same application context boundary.

Keep the existing context value shape intact and move the force-reload polling state with the provider because doesApplicationNeedForceReload is part of that value.
